+++ action/action-simulator/src/main/java/com/powsybl/action/simulator/loadflow/ParallelLoadFlowActionSimulator.java
@@ -12,8 +12,8 @@ import com.powsybl.commons.exceptions.UncheckedInterruptedException;
 import com.powsybl.computation.*;
 import com.powsybl.iidm.network.Network;
 import com.powsybl.iidm.xml.NetworkXml;
-import com.powsybl.security.SecurityAnalysisResult;
-import com.powsybl.security.SecurityAnalysisResultMerger;
+import com.powsybl.security.results.SecurityAnalysisResult;
+import com.powsybl.security.results.SecurityAnalysisResultMerger;
 import com.powsybl.security.json.SecurityAnalysisResultDeserializer;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
