+++ action/action-simulator/src/main/java/com/powsybl/action/simulator/tools/AbstractSecurityAnalysisResultBuilder.java
@@ -10,7 +10,9 @@ import com.powsybl.action.simulator.loadflow.DefaultLoadFlowActionSimulatorObser
 import com.powsybl.action.simulator.loadflow.RunningContext;
 import com.powsybl.contingency.Contingency;
 import com.powsybl.security.*;
+import com.powsybl.security.results.LimitViolationsResult;
 import com.powsybl.security.results.PostContingencyResult;
+import com.powsybl.security.results.SecurityAnalysisResult;
 
 import java.util.*;
 import java.util.stream.Collectors;
