+++ security-analysis/security-analysis-api/src/main/java/com/powsybl/security/SecurityAnalysisProvider.java
@@ -13,6 +13,7 @@ import com.powsybl.contingency.ContingenciesProvider;
 import com.powsybl.iidm.network.Network;
 import com.powsybl.security.interceptors.SecurityAnalysisInterceptor;
 import com.powsybl.security.monitor.StateMonitor;
+import com.powsybl.security.results.SecurityAnalysisResult;
 
 import java.util.Collections;
 import java.util.List;
