+++ security-analysis/security-analysis-api/src/main/java/com/powsybl/security/execution/SecurityAnalysisExecutionImpl.java
@@ -8,6 +8,7 @@ package com.powsybl.security.execution;
 
 import com.powsybl.computation.ComputationManager;
 import com.powsybl.security.*;
+import com.powsybl.security.results.SecurityAnalysisResult;
 
 import java.util.ArrayList;
 import java.util.concurrent.CompletableFuture;
