+++ security-analysis/security-analysis-api/src/main/java/com/powsybl/security/interceptors/CurrentLimitViolationInterceptor.java
@@ -9,7 +9,7 @@ package com.powsybl.security.interceptors;
 import com.powsybl.iidm.network.Branch;
 import com.powsybl.security.LimitViolation;
 import com.powsybl.security.LimitViolationType;
-import com.powsybl.security.LimitViolationsResult;
+import com.powsybl.security.results.LimitViolationsResult;
 import com.powsybl.security.results.PostContingencyResult;
 import com.powsybl.security.extensions.ActivePowerExtension;
 import com.powsybl.security.extensions.CurrentExtension;
