+++ security-analysis/security-analysis-api/src/main/java/com/powsybl/security/interceptors/SecurityAnalysisInterceptor.java
@@ -8,9 +8,9 @@ package com.powsybl.security.interceptors;
 
 import com.powsybl.contingency.Contingency;
 import com.powsybl.security.LimitViolation;
-import com.powsybl.security.LimitViolationsResult;
+import com.powsybl.security.results.LimitViolationsResult;
 import com.powsybl.security.results.PostContingencyResult;
-import com.powsybl.security.SecurityAnalysisResult;
+import com.powsybl.security.results.SecurityAnalysisResult;
 
 /**
  *
