+++ security-analysis/security-analysis-api/src/main/java/com/powsybl/security/json/LimitViolationsResultSerializer.java
@@ -9,7 +9,7 @@ package com.powsybl.security.json;
 import com.fasterxml.jackson.core.JsonGenerator;
 import com.fasterxml.jackson.databind.SerializerProvider;
 import com.fasterxml.jackson.databind.ser.std.StdSerializer;
-import com.powsybl.security.LimitViolationsResult;
+import com.powsybl.security.results.LimitViolationsResult;
 
 import java.io.IOException;
