+++ security-analysis/security-analysis-api/src/main/java/com/powsybl/security/json/PostContingencyResultDeserializer.java
@@ -12,7 +12,7 @@ import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.DeserializationContext;
 import com.fasterxml.jackson.databind.deser.std.StdDeserializer;
 import com.powsybl.contingency.Contingency;
-import com.powsybl.security.LimitViolationsResult;
+import com.powsybl.security.results.LimitViolationsResult;
 import com.powsybl.security.results.PostContingencyResult;
 import com.powsybl.security.results.BranchResult;
 import com.powsybl.security.results.BusResults;
