+++ security-analysis/security-analysis-api/src/main/java/com/powsybl/security/results/LimitViolationsResult.java
@@ -4,7 +4,9 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.security;
+package com.powsybl.security.results;
+
+import com.powsybl.security.LimitViolation;
 
 import java.util.Collections;
 import java.util.List;
