+++ security-analysis/security-analysis-api/src/main/java/com/powsybl/security/results/PreContingencyResult.java
@@ -6,8 +6,6 @@
  */
 package com.powsybl.security.results;
 
-import com.powsybl.security.LimitViolationsResult;
-
 import java.util.*;
 
 /**
