+++ security-analysis/security-analysis-api/src/main/java/com/powsybl/security/results/SecurityAnalysisResult.java
@@ -4,10 +4,10 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.security;
+package com.powsybl.security.results;
 
 import com.powsybl.commons.extensions.AbstractExtendable;
-import com.powsybl.security.results.*;
+import com.powsybl.security.NetworkMetadata;
 
 import java.util.Collections;
 import java.util.List;
