+++ security-analysis/security-analysis-api/src/test/java/com/powsybl/security/SecurityTest.java
@@ -13,7 +13,9 @@ import com.powsybl.contingency.Contingency;
 import com.powsybl.iidm.network.Branch;
 import com.powsybl.iidm.network.Network;
 import com.powsybl.iidm.network.test.EurostagTutorialExample1Factory;
+import com.powsybl.security.results.LimitViolationsResult;
 import com.powsybl.security.results.PostContingencyResult;
+import com.powsybl.security.results.SecurityAnalysisResult;
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.Mockito;
