+++ security-analysis/security-analysis-default/src/test/java/com/powsybl/security/impl/interceptors/SecurityAnalysisInterceptorMock.java
@@ -9,7 +9,9 @@ package com.powsybl.security.impl.interceptors;
 import com.powsybl.security.*;
 import com.powsybl.security.interceptors.DefaultSecurityAnalysisInterceptor;
 import com.powsybl.security.interceptors.SecurityAnalysisResultContext;
+import com.powsybl.security.results.LimitViolationsResult;
 import com.powsybl.security.results.PostContingencyResult;
+import com.powsybl.security.results.SecurityAnalysisResult;
 
 import static org.junit.Assert.*;
